Move `UserContextProvider` to general components directory
