+++ assets/src/onboarding-wizard/index.js
@@ -33,10 +33,10 @@ import { ReaderThemesContextProvider } from '../components/reader-themes-context
 import { ErrorBoundary } from '../components/error-boundary';
 import { ErrorContextProvider } from '../components/error-context-provider';
 import { ErrorScreen } from '../components/error-screen';
+import { UserContextProvider } from '../components/user-context-provider';
 import { PAGES } from './pages';
 import { SetupWizard } from './setup-wizard';
 import { NavigationContextProvider } from './components/navigation-context-provider';
-import { UserContextProvider } from './components/user-context-provider';
 import { SiteScanContextProvider } from './components/site-scan-context-provider';
 import { TemplateModeOverrideContextProvider } from './components/template-mode-override-context-provider';
