+++ assets/src/onboarding-wizard/pages/template-mode/index.js
@@ -8,7 +8,7 @@ import { useEffect, useContext } from '@wordpress/element';
 import { Navigation } from '../../components/navigation-context-provider';
 import { SiteScan } from '../../components/site-scan-context-provider';
 import { ReaderThemes } from '../../../components/reader-themes-context-provider';
-import { User } from '../../components/user-context-provider';
+import { User } from '../../../components/user-context-provider';
 import { Options } from '../../../components/options-context-provider';
 import { Loading } from '../../../components/loading';
 import { TemplateModeOverride } from '../../components/template-mode-override-context-provider';
